+++ lib/togls.rb
@@ -20,12 +20,12 @@ require 'togls/null_toggle'
 require 'togls/rule'
 require 'togls/rules'
 require 'logger'
-require 'togls/release_toggle_registry_manager'
+require 'togls/feature_toggle_registry_manager'
 
 # Togls
 #
 # Togls is the primary interface to the out of the box toggle registry. It is
 # the namespace the DSL is exposed under.
 module Togls
-  include ReleaseToggleRegistryManager
+  include FeatureToggleRegistryManager
 end