+++ lib/togls/feature_toggle_registry.rb
@@ -7,27 +7,17 @@ module Togls
   # respective entities. This plays a significant portion in the primary DSL as
   # well.
   class FeatureToggleRegistry
-    def initialize
+    def initialize(feature_repository)
       @toggle_repository_drivers = [
         Togls::ToggleRepositoryDrivers::InMemoryDriver.new,
         Togls::ToggleRepositoryDrivers::EnvOverrideDriver.new]
-      @feature_repository_drivers =
-        [Togls::FeatureRepositoryDrivers::InMemoryDriver.new]
       @rule_repository_drivers =
         [Togls::RuleRepositoryDrivers::InMemoryDriver.new]
-      @feature_repository = Togls::FeatureRepository.new(
-        @feature_repository_drivers)
       @rule_repository = Togls::RuleRepository.new(@rule_repository_drivers)
       @toggle_repository = Togls::ToggleRepository.new(
-        @toggle_repository_drivers, @feature_repository, @rule_repository)
+        @toggle_repository_drivers, feature_repository, @rule_repository)
       @rule_repository.store(Togls::Rules::Boolean.new(true))
       @rule_repository.store(Togls::Rules::Boolean.new(false))
-    end
-
-    def self.create(&block)
-      feature_toggle_registry = new
-      feature_toggle_registry.instance_eval(&block)
-      feature_toggle_registry
     end
 
     def expand(&block)