+++ spec/togls/test_toggle_registry_spec.rb
@@ -128,37 +128,9 @@ describe Togls::TestToggleRegistry do
     end
   end
 
-  describe ".create" do
-    subject { Togls::TestToggleRegistry }
-
-    it "creates a new instance of a feature toggle registry" do
-      registry = double('registry')
-      expect(subject).to receive(:new).and_return(registry)
-      b = Proc.new {}
-      subject.create(&b)
-    end
-
-    context "when block given" do
-      it "calls instance eval with the passed block" do
-        registry = double('registry')
-        allow(subject).to receive(:new).and_return(registry)
-        b = Proc.new {}
-        expect(registry).to receive(:instance_eval).and_yield(&b)
-        subject.create(&b)
-      end
-    end
-
-    it "returns a configured feature registry object" do
-      b = Proc.new {}
-      expect(subject.create(&b)).to be_a(subject)
-    end
-  end
-
   describe "#expand" do
     it "instance evals the provided block" do
-      registry = Togls::TestToggleRegistry.create do
-        feature(:foo, "some description").on
-      end
+      registry = Togls::TestToggleRegistry.new
 
       expect(registry).to receive(:instance_eval)
       registry.expand do
@@ -167,9 +139,7 @@ describe Togls::TestToggleRegistry do
     end
 
     it "returns the feature toggle repository" do
-      registry = Togls::TestToggleRegistry.create do
-        feature(:foo, "some description").on
-      end
+      registry = Togls::TestToggleRegistry.new
 
       block = Proc.new {}
       expect(registry.expand(&block)).to eq(registry)